Revert "Bug 1966196 - Expand EnvironmentCoordinate hops bitwidth to 16 r=nbp" for causing spidermonkey bustages.
This reverts commit f0c591856b1421d38a70231f3912c494309689e4.

UltraBlame original commit: 110f199bac6f9f76509cd993fb82a6aaa9854a6a

+++ js/src/frontend/EmitterScope.h
@@ -57,7 +57,7 @@ class MOZ_STACK_CLASS EmitterScope : public Nestable<EmitterScope> {
 #endif
 
   
-  uint16_t environmentChainLength_;
+  uint8_t environmentChainLength_;
